requirements.txt
# Core dependencies for textgrad-classifier
textgrad>=0.1.4
litellm>=1.33.6
openai>=1.0.0
typer>=0.12.3
faker>=25.2
pandas>=2.2
scikit-learn>=1.5
requests>=2.32
pyyaml>=6.0
numpy>=1.24.0
tqdm>=4.65.0
python-dotenv>=1.0.0

# Optional: for notebook execution in CI
papermill>=2.5.0
jupyter>=1.0.0